Hide scrollbars
It looks as though react-data-gris is designed to be used as a fixed/constrained width element within the viewport boundary, using a canvas to provide scrolling up/down the grid.
With large datasets this has the consequence that we can get 2 sets of scroll bars - one for the canvas and the other for the viewport.
Try suppressing the canvas scrollbars.

However, this has the effect of breaking the positioning of the editor cell, which used position: fixed, which is tied to the viewport.
Try using position: absolute to deal with this.